backend/requirements.txt
requests>=2.28.0
python-docx>=0.8.11
PyPDF2>=2.11.1
